models data

+++ framework/app/Http/Controllers/UserController.php
@@ -1,22 +1,30 @@
 <?php
 
 namespace App\Http\Controllers;
+
+use App\Hospital;
 use App\Http\Requests\UserCreateRequest;
 use App\Http\Requests\UserUpdateRequest;
 use App\Permission;
 use App\Role;
-use App\Models\User;
+use App\User;
 use Illuminate\Http\Request;
-
-class UserController extends Controller {
+use App\Trait\PermissionModule;
+use DB;
+class UserController extends Controller
+{
+	use PermissionModule;
 	/**
 	 * Display a listing of the resource.
 	 *
 	 * @return \Illuminate\Http\Response
 	 */
 
-	public function index() {
+	public function index()
+	{
 		$this->availibility('View Users');
+		
+
 		$index['page'] = 'users';
 		$index['users'] = User::all();
 
@@ -28,11 +36,12 @@ class UserController extends Controller {
 	 *
 	 * @return \Illuminate\Http\Response
 	 */
-	public function create() {
+	public function create()
+	{
+		$this->availibility('Create Users');
 		$index['page'] = 'users';
 		$index['roles'] = Role::all();
-
-		$this->availibility('Create Users');
+		$index['hospitals'] = Hospital::all();
 		return view('users.create', $index);
 	}
 
@@ -42,15 +51,32 @@ class UserController extends Controller {
 	 * @param  \Illuminate\Http\Request  $request
 	 * @return \Illuminate\Http\Response
 	 */
-	public function store(UserCreateRequest $request) {
-		$user = User::create($request->only('name', 'phone', 'email'));
+	public function store(UserCreateRequest $request)
+	{
+		$user = DB::table('users')->where('email',$request->email)->first();
+		if(!is_null($user)){
+			return redirect()->back()->with('flash_message', 'User before same email Alreay Exist Please Contact super adminstator for further query.')->withInput();
+		}
+		// dd($user);
+		$user = User::create($request->only('name', 'email'));
 		$user->password = bcrypt($request->password);
 		$user->assignRole($request->role);
+		$role = Role::find($request->role);
+		$r_p = $role->permissions->pluck('name')->toArray();
+		$user->syncPermissions($r_p);
+		if ($request->select_all == 1) {
+			$user->select_all = 1;
+		} else {
+			$user->select_all = 0;
+		}
 		$user->save();
-
+		$user->hospitals()->sync($request->hospitals);
+		// dd($user);
 		return redirect()->route('users.index')
-			->with('flash_message',
-				'User "' . $user->name . '" added.');
+			->with(
+				'flash_message',
+				'User "' . $user->name . '" added.'
+			);
 	}
 
 	/**
@@ -59,16 +85,22 @@ class UserController extends Controller {
 	 * @param  int  $id
 	 * @return \Illuminate\Http\Response
 	 */
-	public function edit($id) {
+	public function edit($id)
+	{
+		$this->availibility('Edit Users');
+		$index['permission_array'] = $this->module()['permission_array'];
+		$index['module_names'] = $this->module()['module_names'];
 		$index['page'] = 'users';
 		$index['user'] = User::findOrFail($id);
 		$index['roles'] = Role::all();
-		$this->availibility('Edit Users');
+		$index['hospitals'] = Hospital::all();
 		$permissions = Permission::all();
-		$role_p = $index['user']->getPermissionsViaRoles()->pluck('id')->toArray();
-
-		$index['permissions'] = Permission::whereNotIn('id', $role_p)
-			->get();
+		$index['selectedHospitalIds'] = $index['user']->hospitals->pluck('id')->toArray();
+		// $role_p = $index['user']->getPermissionsViaRoles()->pluck('id')->toArray();
+		//  dd(\Auth::user()->givePermissionTo(Permission::get()));  
+		// $index['permissions'] = Permission::whereNotIn('id', $role_p)
+		// 	->get();
+		$index['permissions'] = Permission::get();
 
 		return view('users.edit', $index);
 	}
@@ -80,30 +112,57 @@ class UserController extends Controller {
 	 * @param  int  $id
 	 * @return \Illuminate\Http\Response
 	 */
-	public function update(UserUpdateRequest $request, $id) {
+	public function update(UserUpdateRequest $request, $id)
+	{
+		$user = DB::table('users')->where('email', $request->email)->where('id','!=',$id)->first();
+		if (!is_null($user)) {
+			return redirect()->back()->with('flash_message', 'User before same email Alreay Exist Please Contact super adminstator for further query.')->withInput();
+		}
+		// dd($request->all());
 		$user = User::findOrFail($id);
-		$user->update($request->only('name', 'phone','email'));
-		$user->password = bcrypt($request->password);
+		$user->update($request->only('name', 'email'));
+		if ($request->password != null) {
+
+			$user->password = bcrypt($request->password);
+		}
+
+		// $user->password = bcrypt($request->password);
+
 		if ($request->role) {
 			$user->roles()->sync($request->role);
 		} else {
 			$user->roles()->detach();
 		}
+
+		// if ($request->role != $user->role_id) {
+		// 	$role = Role::find($request->role);
+		// 	$r_p = $role->permissions->pluck('name')->toArray();
+		// 	$user->syncPermissions($r_p);
+		// } 
+		// else {
+			if (isset($request->permissions)) {
+				$perm = Permission::whereIn('name', $request->permissions)->get();
+				$user->syncPermissions($perm->pluck('name')->toArray());
+				// dd($perm->pluck('name'));
+			} else {
+				$user->revokePermissionTo(Permission::pluck('name')->toArray());
+			}
+		// }
 		$user->role_id = $request->role;
-		if (isset($request->permissions)) {
-			$perm = Permission::whereIn('id', $request->permissions)->get();
-			$user->syncPermissions($perm->pluck('name')->toArray());
+		if ($request->select_all == 1) {
+			$user->select_all = 1;
 		} else {
-			$user->revokePermissionTo(Permission::pluck('name')->toArray());
+			$user->select_all = 0;
 		}
-
 		$user->save();
-
+		$user->hospitals()->sync($request->hospitals);
 		app()['cache']->forget('spatie.permission.cache');
 
 		return redirect()->route('users.index')
-			->with('flash_message',
-				'User "' . $user->name . '" updated.');
+			->with(
+				'flash_message',
+				'User "' . $user->name . '" updated.'
+			);
 	}
 
 	/**
@@ -112,26 +171,40 @@ class UserController extends Controller {
 	 * @param  int  $id
 	 * @return \Illuminate\Http\Response
 	 */
-	public function destroy($id) {
+	public function destroy($id)
+	{
 		$user = User::findOrFail($id);
 		$this->availibility('Delete Users');
 
 		if ($user->id == \Auth::user()->id) {
-			return redirect('admin/users')->with('flash_message_error',
-				' Logged In User  can not Delete.');
+			return redirect('admin/users')->with(
+				'flash_message_error',
+				' Logged In User  can not Delete.'
+			);
 		}
 		$user->delete();
-		return redirect('admin/users')->with('flash_message',
-			'User "' . $user->name . '" deleted.');
+		return redirect('admin/users')->with(
+			'flash_message',
+			'User "' . $user->name . '" deleted.'
+		);
 	}
-	public static function availibility($method) {
-		$r_p = \Auth::user()->getPermissionsViaRoles()->pluck('name')->toArray();
-		if (\Auth::user()->hasPermissionTo($method)) {
+	public static function availibility($method)
+	{
+		// $r_p = \Auth::user()->getPermissionsViaRoles()->pluck('name')->toArray();
+		if (\Auth::user()->hasDirectPermission($method)) {
 			return true;
-		} elseif (!in_array($method, $r_p)) {
-			abort('401');
 		} else {
-			return true;
+			abort('401');
 		}
+		// elseif (!in_array($method, $r_p)) {
+		// 	abort('401');
+		// }
+	}
+	public function user_permissions(Request $request){
+		$index['data'] = User::find($request->user_id);
+		$index['permission_array'] = $this->module()['permission_array'];
+		$index['module_names'] = $this->module()['module_names'];
+		$index['view'] = view('user_role_permission.table_edit_user')->with($index)->render();
+		return response()->json($index);
 	}
 }
